requirements.txt
bandit==1.7.0
coverage==5.5
Django==3.1.7
django-autocomplete-light==3.8.2
django-bootstrap4==2.3.1
django-crispy-forms==1.11.2
django-extensions==3.1.1
django-multiselectfield==0.1.12
django-recaptcha==2.0.6
django-robots==4.0
django-allauth==0.45.0
djangorestframework==3.12.4
flake8==3.9.0
pyflakes==2.3.1
pylint==2.7.2
pylint-django==2.4.2
python-decouple==3.4
pytz==2021.1
six==1.15.0
smmap2==3.0.1
django-sendgrid-v5==1.1.1
